requirements.txt
flask
psycopg2-binary
python-dotenv
twilio
groq
google-generativeai
requests
gunicorn
